Fix various abs() issues.
The abs() function should take an int as argument, but there were
several instances in the code where it was taking an unsigned int or a
double.  In these case, we took one of the following approaches:

* If the argument was a double, use fabs() instead.
* If the argument was unsigned and was certainly going to be positive
  (i.e,. no subtraction), then drop abs() altogether.
* If the argument was unsigned as result of adding or subtracting signed
  and unsigned ints, then we cast all the unsigned ints to signed ints.

Original patch from Doug Torrance <dtorrance@piedmont.edu> for wmaker,
updated to fit with awmaker.

+++ src/menu.c
@@ -1422,7 +1422,7 @@ static void getScrollAmount(WMenu *menu, int *hamount, int *vamount)
 		*hamount = WMIN(MENU_SCROLL_STEP, abs(menuX1));
 	} else if (xroot >= (rect.pos.x + rect.size.width - 2) && menuX2 > (rect.pos.x + rect.size.width - 1)) {
 		/* scroll to the left */
-		*hamount = WMIN(MENU_SCROLL_STEP, abs(menuX2 - rect.pos.x - rect.size.width - 1));
+		*hamount = WMIN(MENU_SCROLL_STEP, abs(menuX2 - rect.pos.x - (int) rect.size.width - 1));
 		if (*hamount == 0)
 			*hamount = 1;
 
@@ -1434,7 +1434,7 @@ static void getScrollAmount(WMenu *menu, int *hamount, int *vamount)
 		*vamount = WMIN(MENU_SCROLL_STEP, abs(menuY1));
 	} else if (yroot >= (rect.pos.y + rect.size.height - 2) && menuY2 > (rect.pos.y + rect.size.height - 1)) {
 		/* scroll up */
-		*vamount = WMIN(MENU_SCROLL_STEP, abs(menuY2 - rect.pos.y - rect.size.height - 2));
+		*vamount = WMIN(MENU_SCROLL_STEP, abs(menuY2 - rect.pos.y - (int) rect.size.height - 2));
 		*vamount = -*vamount;
 	}
 }
